==> 02-DataManip.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---
# Data Manipulation {#Start2}

```{r tidyr2, include = FALSE, warning = FALSE}

library(knitr)
opts_chunk$set(tidy.opts=list(width.cutoff=50), tidy = FALSE)

```

In this Chapter, we will import the raw data and wrangle it into a format suitable for creating summary statistics. 

## Installing packages {#Start2.1}

The functionality of some packages may require updated versions of [R](https://www.r-project.org/) and [RStudio](https://www.rstudio.com). To avoid errors, please ensure you are using the most recent releases of R and RStudio, and update your R packages.
```{r update, eval = FALSE}
update.packages()       

# may need to run this as an R administrator if you run into 'permission denied' issues
```

You need to load the package each time you open a new R session. 
```{r load packages, eval = FALSE, message=FALSE}
#If you need to install the package, use the following script before loading the package. You only need to install the package once on your computer, and then it will always be available. 
#For example: 
#install.package("tidyverse")

require(tidyverse)
require(lubridate)
require(data.table)
require(reshape)
require(tibble)

#If you need to install `naturecounts` this can be done from GitHub with the remotes package.
#install.packages("remotes")
#remotes::install_github("BirdStudiesCanada/naturecounts")

require(naturecounts)
```

## Working Directory {#Start2.2}

First, lets set up your working directory. If you don't know what your current working directory is you can check with the following script. 
```{r directory, eval = FALSE, message=FALSE}
getwd()
``` 

If you want to save your data in a particular location on your computer, set your directory with the following script. 
```{r directory set, eval = FALSE, message=FALSE}
#This is just an example that will need to be modified to the path on your computer

setwd("C:/Users/username/Documents/foldername/")
```

Alternatively, you can set the working directory using the `Session` tab > `Set Working Directory` > `Choose Directory` 

Now we will a create folder to save the raw data and store any outputs. We will also set shortcuts to these folders for future use. 
```{r create folder, eval = FALSE, warning=FALSE}
dir.create("Data")
dir.create("Output")

out.dir <- paste("Output/")
dat.dir <- paste("Data/")
```

## Load Data {#Start2.3}

To access data using the naturecounts R package, you must [sign up](https://www.birdscanada.org/birdmon/default/register.jsp) for a **free** account. Further, if you would like to access Level 3 or 4 collections you must make a [data request](https://www.birdscanada.org/birdmon/default/searchquery.jsp). For additional details and step-by-step instructions, we encourage you to review [NatureCounts: An Introductory Tutorial](https://birdstudiescanada.github.io/NatureCounts_IntroTutorial/).

There are two `collections` in NatureCounts that are associates with the Nightjar program.
```{r collections, eval=FALSE, message=FALSE}
# Creates a table of available collections in naturecounts. 

collections<-meta_collections()

```

WRNIGHTJAR = Wildlife Research Nightjar survey (2010-2015)	
NIGHTJAR = Canadian Nightjar Survey (2016-current)

Although the WildResearch Nightjar Survey dataset spans the period 2010 to 2059, this contains several different data collection protocols and the current national protocol has only been in place since 2016 . We will therefore use the latter of these datasets for the following analyses. 

> The download will not work unless you replace `"user"` with your personal user name, and in the second example provide the `request_id` that will be sent to you once you have permission to access the data. You will be prompted to enter your password.  

```{r data, eval=FALSE, message=FALSE}

data<-nc_data_dl(collections=	"NIGHTJAR", fields_set = "extended", username = "user", info="reason for downloading data")

#or

data <- nc_data_dl(request_id = 000000, fields_set = "extended", username = "user", info="reason for downloading data")
```

Save the raw data in your `Data` directory for future use.
```{r savedata, eval=FALSE, message=FALSE}

write.csv(data, "Data/Raw_Nightjar.csv")

# To load the data from the Data directory
# data<-read.csv("Data/Raw_Nightjar.csv")

```

## Select Data {#Start2.4}

Select the column we want for the summaries stats and analysis. 
```{r select, eval=FALSE, message=FALSE}

dat<-data %>% select(SamplingEventIdentifier, SurveyAreaIdentifier, Locality, SiteCode, utm_square, bcr, survey_year, survey_month, survey_day, TimeCollected, CollectorNumber, CommonName, SpeciesCode, ObservationDescriptor, ObservationCount, ObservationDescriptor2, ObservationCount2, ObservationDescriptor3, ObservationCount3, ObservationDescriptor4, ObservationCount4, ObservationDescriptor5, ObservationCount5, ObservationDescriptor6, ObservationCount6, ObservationDescriptor7, ObservationCount7,
ObservationDescriptor8, ObservationCount8,
ObservationDescriptor9, ObservationCount9,
EffortMeasurement2, EffortUnits2, 
EffortMeasurement3, EffortUnits3, 
EffortMeasurement4, EffortUnits4, 
EffortMeasurement5, EffortUnits5, 
EffortMeasurement6, EffortUnits6, 
EffortMeasurement7, EffortUnits7, 
EffortMeasurement8, EffortUnits8) %>% 
filter(SiteCode != "#N/A")

```

##Filter {#Start2.5}

We will assign Province based on the `SiteCode`. 
```{r sep, eval=FALSE, message=FALSE}

dat<-dat %>% separate(SiteCode, c("Prov", "Site"), sep = "(?<=[A-Za-z])(?=[0-9])", remove=FALSE)
```

Filter out data points that are not useful or missing information
```{r filter}

dat<-dat %>% filter(survey_year!=2015) %>% filter(Prov!="NJ")

```

Save filtered data for future use
```{r save, message=FALSE, eval=FALSE}

write.csv(dat, "Data/Filter_Nightjar.csv")

```
